Remove beta versions of browsers for CI tests
These were included to test for regressions with upcoming versions of
browsers. On the Sauce Labs test platform these tests turn out to be flaky
and error out without explanation, however. Perhaps more generally, test
success with beta software is too unreliable to be used for gating. Since this
is an intended use of the CI test suite, stop using beta software there.